Don't Panic on PKCS11 errors
The plugin should not panic on PKCS11 errors, since they are usually
caused by invalid configuration.

In particular, if the PIN is incorrect, repeated restarts result in
repeated attempts with the same incorrect PIN, this may result in the
token being locked or wiped.

Sleep forever if the PIN is incorrect

Change-Id: Icb2b9320c42b6728049f09265695cac0183ce7e4